Fixed the lightning gun 'target bounces' bug. This was in the original quakec for customtf, but apparently never was being evaluated correctly before someone fixed it for quakeforge aaaaandd.... IMHO, people shouldn't bounce into the air like they're on trampolines when you hit them with the lightning gun :)
Also edited the way excess cells are discharged. No longer is there a cap on the minimum amount of cells that can be evaluted.
Original code made it so that if you had 101 cells and discharged it was the same as if you had 175... IMHO, not funny :)

Also, the excess cells output power is no longer dropped by a third of normal - I may change this back if there are problems with the lg
being too easy to kill people with excess cells... But then again, by nature excess cells are a one time shot - if you miss, they're gone.

Tim McGrath (Misty)